fixed a bug on profile page
