+++ frontend/src/App.jsx
@@ -8,7 +8,7 @@ import NotFoundPage from "./pages/NotFound";
 import UserContext from "./contexts/current-user-context";
 import { checkForLoggedInUser } from "./adapters/auth-adapter";
 import UsersPage from "./pages/Users";
-import UserPage from "./pages/User";
+import UserPage from "./pages/Profile";
 import Feed from "./pages/Feed";
 import MyReports from "./pages/MyReports";
 import MakeAReport from "./pages/MakeAReport";
@@ -42,7 +42,7 @@ export default function App() {
           <Route path="/feed" element={<Feed />} />
           <Route path="/reports-log" element={<MyReports />} />
           <Route path="/report" element={<MakeAReport />} />
-          <Route path="/users/:user_id" element={<Profile />} />
+          <Route path="/users/:id" element={<Profile />} />
 
           {/* Fallback Path */}
           <Route path="*" element={<NotFoundPage />} />
